requirements.txt
python-telegram-bot
flask-sqlalchemy
psycopg2-binary
flask
flask-migrate